+++ R/createPATimeseriesObject.R
@@ -25,7 +25,8 @@
 #' \dontrun{
 #' initializeMazamaSpatialUtils()
 #' pas <- pas_load()
-#' pat_raw <- downloadParseTimeseriesData(pas, name ='North Bend Weather', startdate = 20181908)
+#' pat_raw <- downloadParseTimeseriesData(pas, label = 'North Bend Weather', 
+#'                                        startdate = 20180908)
 #' pat <- createPATimeseriesObject(pat_raw)
 #' }