+++ R/pat_outliers.R
@@ -76,18 +76,18 @@ pat_outliers <- function(
   
   if ( FALSE ) {
     
-    windowSize = 23
-    thresholdMin = 8
-    replace = FALSE
-    showPlot = TRUE
-    # data_shape = 18 
-    # data_size = 1
-    # data_color = "black"
-    # data_alpha = 0.5
-    # outlier_shape = 8 
-    # outlier_size = 1
-    # outlier_color = "red"
-    # outlier_alpha = 1.0
+    windowSize <- 23
+    thresholdMin <- 8
+    replace <- FALSE
+    showPlot <- TRUE
+    # data_shape <- 18 
+    # data_size <- 1
+    # data_color <- "black"
+    # data_alpha <- 0.5
+    # outlier_shape <- 8 
+    # outlier_size <- 1
+    # outlier_color <- "red"
+    # outlier_alpha <- 1.0
      
   }